Post-0.6.3, bump version number to 0.6.4-prerelease
We're doing this so that sites that are running a prerelease build of 0.6.4 will have a different version number from sites rnning a stable 0.6.3 build. The `-prerelease` in the version number is so that sites running 0.6.4-prerelease will run the Largo update when they're finally updated to 0.6.4.

The `0.6.4-prerelease` string should be treated by [version_compare()](https://www.php.net/manual/en/function.version-compare.php) as less than `0.6.4`, according to the documentation for that function.